--- app/src/main/res/layout/fragment_search_list.xml
@@ -1,66 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:descendantFocusability="blocksDescendants">
-
-    <io.legado.app.ui.widget.recycler.scroller.FastScrollRecyclerView
-        android:id="@+id/recycler_view"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:overScrollMode="never"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toTopOf="@+id/ll_search_base_info" />
-
-    <LinearLayout
-        android:id="@+id/ll_search_base_info"
-        android:layout_width="match_parent"
-        android:layout_height="36dp"
-        android:background="@color/background"
-        android:paddingLeft="10dp"
-        android:paddingRight="10dp"
-        android:elevation="5dp"
-        android:gravity="center_vertical"
-        android:orientation="horizontal"
-        app:layout_constraintBottom_toBottomOf="parent">
-
-        <TextView
-            android:id="@+id/tv_current_search_info"
-            android:layout_width="0dp"
-            android:layout_height="match_parent"
-            android:layout_weight="1"
-            android:background="?android:attr/selectableItemBackgroundBorderless"
-            android:ellipsize="middle"
-            android:paddingLeft="10dp"
-            android:paddingRight="10dp"
-            android:singleLine="true"
-            android:gravity="center_vertical"
-            android:textColor="@color/primaryText"
-            android:textSize="12sp" />
-
-        <androidx.appcompat.widget.AppCompatImageView
-            android:id="@+id/iv_search_content_top"
-            android:layout_width="36dp"
-            android:layout_height="match_parent"
-            android:background="?android:attr/selectableItemBackgroundBorderless"
-            android:contentDescription="@string/go_to_top"
-            android:src="@drawable/ic_arrow_drop_up"
-            android:tooltipText="@string/go_to_top"
-            app:tint="@color/primaryText"
-            tools:ignore="UnusedAttribute" />
-
-        <androidx.appcompat.widget.AppCompatImageView
-            android:id="@+id/iv_search_content_bottom"
-            android:layout_width="36dp"
-            android:layout_height="match_parent"
-            android:background="?android:attr/selectableItemBackgroundBorderless"
-            android:contentDescription="@string/go_to_bottom"
-            android:src="@drawable/ic_arrow_drop_down"
-            android:tooltipText="@string/go_to_bottom"
-            app:tint="@color/primaryText"
-            tools:ignore="UnusedAttribute" />
-    </LinearLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
